refactor: split executor.go into focused files to satisfy 600-line limit
Extract workdir provisioning, output processing, utility helpers, and runner
methods from executor.go into dedicated files; executor.go now sits at 563 lines.
